Cambios en escenas
Escenas cambiadas y con mejor diseño.

+++ Snake-Pet/Assets/Scripts/formulario.cs
@@ -35,7 +35,6 @@ public class Usuario
 {
     public string ID_Usuario;
     public string Nombre;
-    public string Password;
     public List<Rutina> Rutinas;
     public List<Mascota> Mascotas;
 }
@@ -49,42 +48,66 @@ public class UsuariosData
 public class formulario : MonoBehaviour
 {
     public TMP_InputField nombreInput;
-    public TMP_InputField passwordInput;
     public Button guardarButton;
+    public GameObject Serpiente; // Prefab del GameObject de la mascota
     private string filePath;
 
     private void Start()
     {
         // Establecer la ruta del archivo JSON en la carpeta de Assets
         filePath = Path.Combine(Application.dataPath, "Scripts/snake.json");
-        
+
         // Verificar si los InputFields están asignados correctamente
-        if (nombreInput == null || passwordInput == null)
+        if (nombreInput == null)
         {
             Debug.LogError("Los campos de entrada no están asignados correctamente en el Inspector.");
             return;
-        }
+        } 
 
         guardarButton.onClick.AddListener(GuardarUsuario);
+    }
+
+    public void Update()
+    {
+        // Validar si el nombreInput está vacío
+        if (string.IsNullOrEmpty(nombreInput.text))
+        {
+            // Si el campo de nombre está vacío, el botón se desactiva
+            guardarButton.interactable = false;
+        } 
+        else 
+        {
+            // Si el campo de nombre tiene algo, el botón se activa
+            guardarButton.interactable = true;
+        }
     }
 
     private void GuardarUsuario()
     {
         string nombre = nombreInput.text;
-        string password = passwordInput.text;
+
+        // Cargar los datos actuales de usuarios
+        UsuariosData usuariosData = CargarUsuarios();
+
+        // Verificar si el nombre de usuario ya existe
+        if (usuariosData.usuarios.Exists(u => u.Nombre == nombre))
+        {
+            guardarButton.interactable = false;
+
+            return;
+        }
 
         Usuario nuevoUsuario = new Usuario
         {
             ID_Usuario = Guid.NewGuid().ToString(),
             Nombre = nombre,
-            Password = password,
             Rutinas = new List<Rutina>(),
             Mascotas = new List<Mascota>()
         };
 
-        UsuariosData usuariosData = CargarUsuarios();
         usuariosData.usuarios.Add(nuevoUsuario);
 
+        // Convertir a JSON y guardar en el archivo
         string jsonData = JsonUtility.ToJson(usuariosData, true);
         File.WriteAllText(filePath, jsonData);
